backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
pgvector==0.3.2
ollama==0.3.3
PyPDF2==3.0.1
scikit-learn==1.5.2
numpy==1.26.4
gunicorn==22.0.0
